pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "HxongResourceDownloader"
version = "0.0.3"
authors = [
  { name="", email="author@example.com" },
]
description = "A simple package download file on telegram"
readme = "README.md"
requires-python = ">=3.11.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "Telethon >= 1.37.0",
    "cryptg >= 0.5.0.post0",
    "pathlib >= 1.0.1",
    "patool >= 3.0.3"
]

[project.urls]
Homepage = "https://github.com/kou843383495/Hxong-Resource-Downloader"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

